+++ Content.Client/Stack/StackSystem.cs
@@ -1,3 +1,4 @@
+using Content.Client.Items;
 using Content.Shared.Stacks;
 using JetBrains.Annotations;
 
@@ -6,6 +7,17 @@ namespace Content.Client.Stack
     [UsedImplicitly]
     public sealed class StackSystem : SharedStackSystem
     {
+        public override void Initialize()
+        {
+            base.Initialize();
+            SubscribeLocalEvent<StackComponent, ItemStatusCollectMessage>(OnItemStatus);
+        }
+
+        private void OnItemStatus(EntityUid uid, StackComponent component, ItemStatusCollectMessage args)
+        {
+            args.Controls.Add(new StackStatusControl(component));
+        }
+
         public override void SetCount(EntityUid uid, int amount, SharedStackComponent? component = null)
         {
             if (!Resolve(uid, ref component))
